agentic-os/backend/services/tools/src/requirements.txt
# For tools:
pyyaml==6.0.2
prompt_toolkit==3.0.50
semgrep==1.114.0
javalang==0.13.0
pathspec==0.12.1
backoff==2.2.1
slack-sdk==3.35.0
slack-bolt==1.23.0

#openai==1.68.2
#marshmallow==3.26.1
#rich==13.5.2
#typer==0.15.2

# For logs:
python-json-logger==3.3.0

# For services:
fastapi==0.115.12
uvicorn==0.34.1
nats-py==2.10.0

# For tests:
pytest==8.3.5
httpx==0.28.1

